Fixing bugs and issues  - Fixed instructions coding  - Fixed imediate addressation  - Fixed negative number issue

+++ ProcessorStructure.h
@@ -3,25 +3,25 @@
 class ProcessorStructure
 {
 public:
-	static int T;
-	static int PC;
-	static int SP;
-	static int RG[16];
-	static int flags; // NZVC
+	static short int T;
+	static unsigned short int PC;
+	static unsigned short int SP;
+	static short int RG[16];
+	static short flags; // NZVC
 	static bool BE0;
 	static bool BE1;
 	static bool BVI;
 	static bool BPO;
 	static int MEM[65536];
- 	static long int MPM[116];
-	static int MAR;
-	static long int MIR;
-	static int ADR;
-	static int MDR;
+ 	static unsigned long long int MPM[116];
+	static unsigned short int MAR;
+	static unsigned long long int MIR;
+	static unsigned short int ADR;
+	static short int MDR;
 	static int IR;
-	static int SBUS;
-	static int DBUS;
-	static int RBUS;
+	static short int SBUS;
+	static short int DBUS;
+	static short int RBUS;
 	static short C;
 	static short N;
 	static short V;
